pyproject.toml
[project]
name = "day06-ai-product-hackathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.6",
    "pandas>=3.0.2",
    "pydantic>=2.12.5",
    "pytest>=9.0.3",
    "python-dotenv>=1.2.2",
]
